fix: refine tasks cli analytics handling
